Refactor line minimap component with simplified styling
- Move utilities to component folder from utilities directory
- Update line styles to use bg-gray12/bg-gray9 classes with 1px width
- Fix indicator positioning with pinned left position (no animation)
- Remove animation and scroll effects for static display
- Update indicator to use bg-orange class and CSS variable
